education/requirements.txt
Django==3.2
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.5
django-embed-video==1.4.0
django-extensions==3.1.3
djangorestframework==3.12.4
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.3.0
django-crispy-forms==1.11.2
Pillow==8.2.0
